Refactor DamagePickerView to separate view
